Re-organizes `stats.py` into a submodule.
