+++ slideflow/__init__.py
@@ -22,7 +22,7 @@ def backend():
     return os.environ['SF_BACKEND']
 
 
-from slideflow import io, model, norm
+from slideflow import io, model, norm, stats
 from slideflow.dataset import Dataset
 from slideflow.heatmap import Heatmap
 from slideflow.model import DatasetFeatures, ModelParams
