--- slideflow/stats.py
@@ -1,1902 +1,0 @@
-import csv
-import multiprocessing as mp
-import os
-import pickle
-import sys
-import time
-from functools import partial
-from multiprocessing.sharedctypes import Value
-from os.path import join
-from random import sample
-from types import SimpleNamespace
-from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union
-
-import numpy as np
-import pandas as pd
-from lifelines.utils import concordance_index as c_index
-from mpl_toolkits.mplot3d import Axes3D
-from scipy import stats
-from sklearn import metrics
-from sklearn.cluster import KMeans
-from sklearn.metrics import pairwise_distances_argmin_min
-from tqdm import tqdm
-from pandas.core.frame import DataFrame
-
-import slideflow as sf
-from slideflow import errors
-from slideflow.util import Path, ProgressBar, as_list
-from slideflow.util import colors as col
-from slideflow.util import log, to_onehot
-
-if TYPE_CHECKING:
-    import neptune.new as neptune
-    import tensorflow as tf
-    import torch
-    from matplotlib.axes import Axes
-    from matplotlib.figure import Figure
-
-    from slideflow.model import DatasetFeatures
-
-# TODO: remove 'hidden_0' reference as this may not be present
-# if the model does not have hidden layers
-
-
-class SlideMap:
-    """Two-dimensional slide map for visualization & backend for mosaic maps.
-
-    Slides are mapped in 2D either explicitly with pre-specified coordinates,
-    or with dimensionality reduction from post-convolutional layer weights,
-    provided from :class:`slideflow.DatasetFeatures`.
-    """
-
-    def __init__(
-        self,
-        slides: List[str],
-        cache: Optional[Path] = None
-    ) -> None:
-        """Backend for mapping slides into two dimensional space. Can use a
-        DatasetFeatures object to map slides according to UMAP of features, or
-        map according to pre-specified coordinates.
-
-        Args:
-            slides (list(str)): List of slide names
-            cache (str, optional): Path to PKL file to cache activations.
-                Defaults to None (caching disabled).
-        """
-
-        self.slides = slides
-        self.cache = cache
-        self.df = None  # type: Optional[DatasetFeatures]
-        self.data = None  # type: DataFrame
-        self.map_meta = {}  # type: Dict[str, Any]
-        if self.cache:
-            self.load_cache()
-
-    @classmethod
-    def from_precalculated(
-        cls,
-        x: Union[np.ndarray, List[int], str],
-        y: Union[np.ndarray, List[int], str],
-        slides: Union[np.ndarray, List[str], str],
-        tfr_index: Union[np.ndarray, List[int], str],
-        data: Optional[DataFrame] = None,
-        cache: Optional[Path] = None
-    ) -> "SlideMap":
-        """Initializes map from precalculated coordinates.
-
-        Args:
-            slides (list(str)): List of slide names.
-            x (list(int)): List of X coordinates for tfrecords.
-            y (list(int)): List of Y coordinates for tfrecords.
-            meta (list(dict)): List of dicts containing metadata for each point
-                 on the map (representing a single tfrecord).
-            labels (list(str)): Labels assigned to each tfrecord, used for
-                coloring TFRecords according to labels.
-            cache (str, optional): Path to PKL file to cache coordinates.
-                Defaults to None (caching disabled).
-        """
-
-        # Read and verify provided input
-        cols = {'x': x, 'y': y, 'slides': slides, 'tfr_index': tfr_index}
-        for col, col_val in cols.items():
-            if isinstance(col_val, str) and data is None:
-                raise ValueError(
-                    f"Could not interpret input {col_val} for arg {col}. "
-                    "Did you mean to supply a DataFrame via 'data'?")
-            elif data is not None:
-                if isinstance(col_val, str) and col_val not in data.columns:
-                    raise ValueError(f"Could not find column {col_val}.")
-                elif isinstance(col_val, str):
-                    cols[col] = data[col_val].values
-            else:
-                cols[col] = col_val
-
-        # Verify lengths of provided input
-        if not all(len(cols[c]) == len(cols['x']) for c in cols):
-            raise ValueError(
-                "Length of x, y, slides, and tfr_index must all be equal."
-            )
-
-        obj_data = pd.DataFrame({
-            'x': pd.Series(cols['x']),
-            'y': pd.Series(cols['y']),
-            'slide': pd.Series(cols['slides']),
-            'tfr_index': pd.Series(cols['tfr_index'])
-        })
-        obj = cls(obj_data.slide.unique())
-        obj.data = obj_data
-        obj.cache = cache
-        obj.save_cache()
-        return obj
-
-    @classmethod
-    def from_features(
-        cls,
-        df: "DatasetFeatures",
-        exclude_slides: Optional[List[str]] = None,
-        recalculate: bool = False,
-        map_slide: Optional[str] = None,
-        cache: Optional[Path] = None,
-        low_memory: bool = False,
-        umap_dim: int = 2,
-        umap: Optional[Any] = None,
-        **umap_kwargs: Any
-    ) -> "SlideMap":
-        """Initializes map from dataset features.
-
-        Args:
-            df (:class:`slideflow.DatasetFeatures`): DatasetFeatures.
-            exclude_slides (list, optional): List of slides to exclude.
-            recalculate (bool, optional):  Force recalculation of umap despite
-                presence of cache.
-            use_centroid (bool, optional): Calculate/map centroid activations.
-            map_slide (str, optional): Either None, 'centroid', or 'average'.
-                If None, will map all tiles from each slide. Defaults to None.
-            cache (str, optional): Path to PKL file to cache coordinates.
-                Defaults to None (caching disabled).
-        """
-
-        if map_slide is not None and map_slide not in ('centroid', 'average'):
-            raise errors.SlideMapError(
-                "map_slide must be None, 'centroid' or 'average', (got "
-                f"{map_slide})"
-            )
-
-        if not exclude_slides:
-            slides = df.slides
-        else:
-            slides = [s for s in df.slides if s not in exclude_slides]
-
-        obj = cls(slides, cache=cache)
-        obj.df = df
-        obj.umap = umap
-        if map_slide:
-            obj._calculate_from_slides(
-                method=map_slide,
-                recalculate=recalculate,
-                low_memory=low_memory,
-                **umap_kwargs
-            )
-        else:
-            obj._calculate_from_tiles(
-                recalculate=recalculate,
-                low_memory=low_memory,
-                dim=umap_dim,
-                **umap_kwargs
-            )
-        return obj
-
-    @property
-    def x(self):
-        return self.data.x.values
-
-    @property
-    def y(self):
-        return self.data.y.values
-
-    def _calculate_from_tiles(
-        self,
-        recalculate: bool = False,
-        **umap_kwargs: Any
-    ) -> None:
-        """Internal function to guide calculation of UMAP from final layer
-        features / activations, as provided by DatasetFeatures.
-
-        Args:
-            recalculate (bool, optional): Recalculate of UMAP despite loading
-                from cache. Defaults to False.
-
-        Keyword Args:
-            dim (int): Number of dimensions for UMAP. Defaults to 2.
-            n_neighbors (int): Number of neighbors for UMAP. Defaults to 50.
-            min_dist (float): Minimum distance for UMAP. Defaults to 0.1.
-            metric (str): UMAP metric. Defaults to 'cosine'.
-            low_memory (bool). Operate UMAP in low memory mode.
-                Defaults to False.
-        """
-        assert self.df is not None
-        if self.data is not None and not recalculate:
-            log.info("Data loaded from cache, will not recalculate")
-
-            # First, filter out slides not included in provided activations
-            self.data = self.data.loc[self.data.slide.isin(self.df.slides)]
-
-            # If UMAP already calculated, update predictions
-            # if prediction filter is provided
-            logits = [
-                self.df.logits[row.slide][row.tfr_index]
-                for _, row in self.data.iterrows()
-            ]
-            predictions = np.argmax(np.array(logits), axis=1)
-            self.data['logits'] = pd.Series(logits)
-            self.data['predictions'] = pd.Series([p for p in predictions])
-            return
-
-        # Calculate UMAP
-        node_activations = np.concatenate([
-            self.df.activations[slide] for slide in self.slides
-        ])
-
-        self.map_meta['num_features'] = self.df.num_features
-        log.info("Calculating UMAP...")
-
-        coordinates = self.umap_transform(node_activations, **umap_kwargs)
-
-        # Assemble dataframe
-        logits = np.concatenate([
-            self.df.logits[slide] for slide in self.slides
-        ])
-        locations = np.concatenate([
-            self.df.locations[slide] for slide in self.slides
-        ])
-        tfrecord_indices = np.concatenate([
-            np.arange(self.df.locations[slide].shape[0])
-            for slide in self.slides
-        ])
-        slides = np.array([
-            slide
-            for slide in self.slides
-            for _ in range(self.df.locations[slide].shape[0])
-        ])
-        data_dict = {
-            'slide': pd.Series(slides),
-            'x': pd.Series(coordinates[:, 0]),
-            'tfr_index': pd.Series(tfrecord_indices),
-            'logits': pd.Series([l for l in logits]).astype(object),
-            'prediction': pd.Series(np.argmax(logits, axis=1)),
-            'location': pd.Series([l for l in locations]).astype(object)
-        }
-        if self.df.hp.uq and self.df.uncertainty != {}:  # type: ignore
-            uncertainty = np.concatenate([
-                self.df.uncertainty[slide] for slide in self.slides
-            ])
-            data_dict.update({
-                'uncertainty': pd.Series(
-                    [u for u in uncertainty]
-                ).astype(object)
-            })
-        if 'dim' not in umap_kwargs or umap_kwargs['dim'] > 1:
-            data_dict.update({
-                'y': pd.Series(coordinates[:, 1]),
-            })
-        self.data = pd.DataFrame(data_dict)
-        self.save_cache()
-
-    def _calculate_from_slides(
-        self,
-        method: str = 'centroid',
-        recalculate: bool = False,
-        **umap_kwargs: Any
-    ) -> None:
-        """ Internal function to guide calculation of UMAP from final layer
-            activations for each tile, as provided via DatasetFeatures, and
-            then map only the centroid tile for each slide.
-
-        Args:
-            method (str, optional): 'centroid' or 'average'. If centroid, will
-                calculate UMAP only from centroid tiles for each slide.
-                If average, will calculate UMAP based on average node
-                activations across all tiles within the slide, then display the
-                centroid tile for each slide.
-            recalculate (bool, optional): Recalculate of UMAP despite loading
-            from cache. Defaults to False.
-            low_memory (bool, optional): Calculate UMAP in low-memory mode.
-                Defaults to False.
-
-        Keyword Args:
-            dim (int): Number of dimensions for UMAP. Defaults to 2.
-            n_neighbors (int): Number of neighbors for UMAP. Defaults to 50.
-            min_dist (float): Minimum distance for UMAP. Defaults to 0.1.
-            metric (str): UMAP metric. Defaults to 'cosine'.
-            low_memory (bool). Operate UMAP in low memory mode.
-                Defaults to False.
-        """
-
-        if method not in ('centroid', 'average'):
-            _m = f'Method must be either "centroid" or "average", not {method}'
-            raise errors.SlideMapError(_m)
-        assert self.df is not None
-
-        # Calculate optimal slide indices and centroid activations
-        log.info("Calculating centroid indices...")
-        opt_idx, centroid_activations = calculate_centroid(self.df.activations)
-
-        # Restrict mosaic to only slides that had enough tiles to calculate
-        # an optimal index from centroid
-        successful_slides = list(opt_idx.keys())
-        num_warned = 0
-        for slide in self.df.slides:
-            if slide not in successful_slides:
-                log.debug(f"No centroid for {col.green(slide)}; skipping")
-        if num_warned:
-            log.warning(f"No centroid for {num_warned} slides.")
-        if self.data is not None and not recalculate:
-            log.info("Slide map loaded from cache.")
-            log.debug("Filtering to include only provided tiles")
-
-            def is_opt(row):
-                return ((row['slide'] in opt_idx)
-                        and (row['tfr_index'] == opt_idx[row['slide']]))
-
-            self.data = self.data.loc[
-                self.data.apply(lambda row : is_opt(row), axis=1)
-            ]
-        else:
-            log.info(f"Calculating UMAP from slide-level {method}...")
-
-            if method == 'centroid':
-                umap_input = np.array([
-                    centroid_activations[slide] for slide in self.slides
-                ])
-            elif method == 'average':
-                umap_input = np.array([
-                    np.mean(self.df.activations[slide], axis=0)
-                    for slide in self.slides
-                ])
-
-            # Calculate UMAP
-            coordinates = self.umap_transform(
-                umap_input,
-                **umap_kwargs
-            )
-
-            # Create dataframe
-            logits = np.stack([
-                self.df.logits[slide][opt_idx[slide]] for slide in self.slides
-            ])
-            locations = np.stack([
-                self.df.locations[slide][opt_idx[slide]] for slide in self.slides
-            ])
-            data_dict = {
-                'slide': pd.Series(self.slides),
-                'x': pd.Series(coordinates[:, 0]),
-                'tfr_index': pd.Series(opt_idx[slide] for slide in self.slides),
-                'logits': pd.Series([l for l in logits]).astype(object),
-                'prediction': pd.Series(np.argmax(logits, axis=1)),
-                'location': pd.Series([l for l in locations]).astype(object)
-            }
-            if self.df.hp.uq and self.df.uncertainty != {}:  # type: ignore
-                uncertainty = np.stack([
-                    self.df.uncertainty[slide][opt_idx[slide]]
-                    for slide in self.slides
-                ])
-                data_dict.update({
-                    'uncertainty': pd.Series(
-                        [u for u in uncertainty]
-                    ).astype(object)
-                })
-            if 'dim' not in umap_kwargs or umap_kwargs['dim'] > 1:
-                data_dict.update({
-                    'y': pd.Series(coordinates[:, 1]),
-                })
-            self.data = pd.DataFrame(data_dict)
-            self.save_cache()
-
-    def activations(self) -> np.ndarray:
-        if self.df is None:
-            raise ValueError(
-                "No associated DatasetFeatures object for reading activations."
-            )
-        return np.array([
-            self.df.activations[row.slide][row.tfr_index]
-            for _, row in self.data.iterrows()
-        ])
-
-    def cluster(self, n_clusters: int) -> None:
-        """Performs clustering on data and adds to metadata labels.
-        Requires a DatasetFeatures backend.
-
-        Clusters are saved to self.data['cluster'].
-
-        Args:
-            n_clusters (int): Number of clusters for K means clustering.
-        """
-
-        if self.df is None:
-            raise errors.SlideMapError(
-                "Unable to cluster; no DatasetFeatures provided"
-            )
-        activations = [
-            self.df.activations[row.slide][row.tfr_index]
-            for _, row in self.data.iterrows()
-        ]
-        log.info(f"Calculating K-means clustering (n={n_clusters})")
-        kmeans = KMeans(n_clusters=n_clusters).fit(activations)
-        labels = kmeans.labels_
-        self.data['cluster'] = labels
-
-    def neighbors(
-        self,
-        slide_categories: Optional[Dict] = None,
-        algorithm: str = 'kd_tree',
-        method: str = 'map',
-        pca_dim: int = 100
-    ) -> None:
-        """Calculates neighbors among tiles in this map, assigning neighboring
-            statistics to tile metadata 'num_unique_neighbors' and
-            'percent_matching_categories'.
-
-        Args:
-            slide_categories (dict, optional): Maps slides to categories.
-                Defaults to None. If provided, will be used to calculate
-                'percent_matching_categories' statistic.
-            algorithm (str, optional): NearestNeighbor algorithm, either
-                'kd_tree', 'ball_tree', or 'brute'. Defaults to 'kd_tree'.
-            method (str, optional): Either 'map', 'pca', or 'features'. How
-                neighbors are determined. If 'map', calculates neighbors based
-                on UMAP coordinates. If 'features', calculates neighbors on the
-                full feature space. If 'pca', reduces features into `pca_dim`
-                space. Defaults to 'map'.
-        """
-        from sklearn.decomposition import PCA
-        from sklearn.neighbors import NearestNeighbors
-        if self.df is None:
-            raise errors.SlideMapError(
-                "Unable perform neighbor search; no DatasetFeatures provided"
-            )
-        log.info(f"Initializing neighbor search (method={method})...")
-        if method == 'map':
-            X = np.stack((self.data.x.values, self.data.y.values), axis=-1)
-        elif method == 'features':
-            X = self.activations()
-        elif method == 'pca':
-            log.info(f"Reducing dimensionality with PCA (dim={pca_dim})...")
-            pca = PCA(n_components=pca_dim)
-            features = self.activations()
-            pca.fit(features)
-            X = pca.transform(features)
-
-        else:
-            raise ValueError(f'Unknown neighbor method {method}.')
-        nbrs = NearestNeighbors(
-            n_neighbors=100,
-            algorithm=algorithm,
-            n_jobs=-1
-        ).fit(X)
-        log.info("Calculating nearest neighbors...")
-        _, indices = nbrs.kneighbors(X)
-
-        def num_category_matching(idx_list, idx):
-            list_cat = np.array([
-                slide_categories[self.data.loc[_i].slide] for _i in idx_list
-            ])
-            idx_cat = slide_categories[self.data.loc[idx].slide]
-            return (list_cat == idx_cat).sum()
-
-        log.info('Matching neighbors...')
-        self.data['num_unique_neighbors'] = [
-            len(self.data.loc[ind].slide.unique())
-            for ind in indices
-        ]
-        if slide_categories:
-            self.data['percent_matching_categories'] = [
-                num_category_matching(ind, i) / len(ind)
-                for i, ind in enumerate(indices)
-            ]
-
-    def filter(self, slides: List[str]) -> None:
-        """Filters map to only show tiles from the given slides.
-
-        Args:
-            slides (list(str)): List of slide names.
-        """
-
-        self.data = self.data.loc[self.data.slide.isin(slides)]
-
-    def umap_transform(
-        self,
-        array: np.ndarray,
-        dim: int = 2,
-        n_neighbors: int = 50,
-        min_dist: float = 0.1,
-        metric: str = 'cosine',
-        **kwargs: Any
-    ) -> np.ndarray:
-        """Generates and returns a umap from a given array, using umap.UMAP"""
-        import umap  # Imported in this function due to long import time
-        if not len(array):
-            raise errors.StatsError("Unable to perform UMAP on empty array.")
-        if self.umap is None:
-            self.umap = umap.UMAP(
-                n_components=dim,
-                verbose=(log.getEffectiveLevel() <= 20),
-                n_neighbors=n_neighbors,
-                min_dist=min_dist,
-                metric=metric,
-                **kwargs
-            )
-        layout = self.umap.fit_transform(array)  # type: ignore
-        return normalize_layout(layout)
-
-    def label_by_uncertainty(self, index: int = 0) -> None:
-        """Labels each point with the tile-level uncertainty, if available.
-
-        Args:
-            index (int, optional): Uncertainty index. Defaults to 0.
-        """
-        if 'label' in self.data.columns:
-            self.data.drop(columns='label')
-        if self.df is None:
-            raise errors.SlideMapError("DatasetFeatures not provided.")
-        if not self.df.hp.uq or self.df.uncertainty == {}:  # type: ignore
-            raise errors.DatasetError(
-                'Unable to label by uncertainty; UQ estimates not available.'
-            )
-        else:
-            uq_labels = np.stack(self.data['uncertainty'].values)[:, index]
-            self.data['label'] = uq_labels
-
-    def label_by_logits(self, index: int) -> None:
-        """Displays each point with label equal to the logits (linear from 0-1)
-
-        Args:
-            index (int): Logit index.
-        """
-        if 'label' in self.data.columns:
-            self.data.drop(columns='label')
-        self.data['label'] = np.stack(self.data['logits'].values)[:, index]
-
-    def label_by_slide(self, slide_labels: Optional[Dict] = None) -> None:
-        """Displays each point as the name of the corresponding slide.
-            If slide_labels is provided, will use this dict to label slides.
-
-        Args:
-            slide_labels (dict, optional): Dict mapping slide names to labels.
-        """
-
-        if 'label' in self.data.columns:
-            self.data.drop(columns='label')
-        if slide_labels:
-            self.data['label'] = self.data.slide.map(slide_labels)
-        else:
-            self.data['label'] = self.data.slide.values
-
-    def label(self, meta: str) -> None:
-        """Displays each point labeled by tile metadata (e.g. 'prediction')
-
-        Args:
-            meta (str): Data column from which to assign labels.
-            translation_dict (dict, optional): If provided, will translate the
-                read metadata through this dictionary.
-        """
-        if 'label' in self.data.columns:
-            self.data.drop(columns='label')
-        self.data['label'] = self.data[meta].values
-
-    def plot(
-        self,
-        subsample: Optional[int] = None,
-        title: Optional[str] = None,
-        cmap: Optional[Dict] = None,
-        xlim: Tuple[float, float] = (-0.05, 1.05),
-        ylim: Tuple[float, float] = (-0.05, 1.05),
-        xlabel: Optional[str] = None,
-        ylabel: Optional[str] = None,
-        legend: Optional[str] = None,
-        ax: Optional["Axes"] = None,
-        categorical: Union[str, bool] = 'auto',
-        **scatter_kwargs: Any
-    ) -> None:
-        """Plots calculated map.
-
-        Args:
-            subsample (int, optional): Subsample to only include this many
-                tiles on plot. Defaults to None.
-            title (str, optional): Title for plot.
-            cmap (dict, optional): Dict mapping labels to colors.
-            xlim (list, optional): List of float indicating limit for x-axis.
-                Defaults to (-0.05, 1.05).
-            ylim (list, optional): List of float indicating limit for y-axis.
-                Defaults to (-0.05, 1.05).
-            xlabel (str, optional): Label for x axis. Defaults to None.
-            ylabel (str, optional): Label for y axis. Defaults to None.
-            legend (str, optional): Title for legend. Defaults to None.
-            ax (matplotlib.axes.Axes, optional): Figure axis. If not supplied,
-                will prepare a new figure axis.
-        """
-        import seaborn as sns
-        from matplotlib import pyplot as plt
-
-        # Make plot
-        if ax is None:
-            fig = plt.figure(figsize=(6, 4.5))
-            ax = fig.add_subplot(111)
-
-        # Subsampling
-        if subsample:
-            plot_df = self.data.sample(subsample)
-        else:
-            plot_df = self.data
-
-        x = plot_df.x
-        y = plot_df.y
-
-        if 'label' in self.data.columns:
-            labels = plot_df.label
-
-            # Check for categorical labels
-            if (categorical is True
-               or not pd.to_numeric(labels, errors='coerce').notnull().all()):
-                log.debug("Interpreting labels as categorical")
-                scatter_kwargs.update(
-                    dict(hue=labels.astype('category'))
-                )
-                unique = list(labels.unique())
-                unique.sort()
-                if len(unique) >= 12:
-                    sns_pal = sns.color_palette("Paired", len(unique))
-                else:
-                    sns_pal = sns.color_palette('hls', len(unique))
-                if cmap is None:
-                    cmap = {unique[i]: sns_pal[i] for i in range(len(unique))}
-            else:
-                log.debug("Interpreting labels as continuous")
-                scatter_kwargs.update(dict(hue=labels))
-
-        umap_2d = sns.scatterplot(
-            x=x,
-            y=y,
-            palette=cmap,
-            ax=ax,
-            **scatter_kwargs
-        )
-        ax.set_ylim(*((None, None) if not ylim else ylim))
-        ax.set_xlim(*((None, None) if not xlim else xlim))
-        if 'hue' in scatter_kwargs:
-            umap_2d.legend(
-                loc='center left',
-                bbox_to_anchor=(1.25, 0.5),
-                ncol=1,
-                title=legend
-            )
-        umap_2d.set(xlabel=xlabel, ylabel=ylabel)
-        if title:
-            ax.set_title(title)
-
-    def plot_3d(
-        self,
-        z: Optional[np.ndarray] = None,
-        feature: Optional[int] = None,
-        subsample: Optional[int] = None,
-        fig: Optional["Figure"] = None,
-    ) -> None:
-        """Saves a plot of a 3D umap, with the 3rd dimension representing
-        values provided by argument "z".
-
-        Args:
-            z (list, optional): Values for z axis. Must supply z or feature.
-                Defaults to None.
-            feature (int, optional): Int, feature to plot on 3rd axis.
-                Must supply z or feature. Defaults to None.
-            subsample (int, optional): Subsample to only include this many
-                tiles on plot. Defaults to None.
-            fig (matplotlib.figure.Figure, optional): Figure. If not supplied,
-                will prepare a new figure.
-        """
-        from matplotlib import pyplot as plt
-
-        if fig is None:
-            fig = plt.figure()
-
-        title = f"UMAP with feature {feature} focus"
-        if self.df is None:
-            raise errors.SlideMapError("DatasetFeatures not provided.")
-        if (z is None) and (feature is None):
-            raise errors.SlideMapError("Must supply either 'z' or 'feature'.")
-
-        # Subsampling
-        if subsample:
-            plot_df = self.data.sample(subsample)
-        else:
-            plot_df = self.data
-
-        # Get feature activations for 3rd dimension
-        if z is None:
-            z = np.array([
-                self.df.activations[row.slide][row.tfr_index][feature]
-                for _, row in plot_df.iterrows()
-            ])
-
-        # Plot tiles on a 3D coordinate space with 2 coordinates from UMAP
-        # and 3rd from the value of the excluded feature
-        ax = Axes3D(fig, auto_add_to_figure=False)
-        fig.add_axes(ax)
-        scatter_kw = dict(c=z, cmap='viridis', linewidth=0.5, edgecolor="k")
-        ax.scatter(plot_df.x, plot_df.y, z, **scatter_kw)
-        ax.set_title(title)
-
-    def save(
-        self,
-        filename: str,
-        dpi: int = 300,
-        **kwargs
-    ):
-        """Save plot of slide map.
-
-        Args:
-            filename (str): File path to save the image.
-            dpi (int, optional): DPI for final image. Defaults to 300.
-
-        Keyword args:
-            subsample (int, optional): Subsample to only include this many
-                tiles on plot. Defaults to None.
-            title (str, optional): Title for plot.
-            cmap (dict, optional): Dict mapping labels to colors.
-            xlim (list, optional): List of float indicating limit for x-axis.
-                Defaults to (-0.05, 1.05).
-            ylim (list, optional): List of float indicating limit for y-axis.
-                Defaults to (-0.05, 1.05).
-            xlabel (str, optional): Label for x axis. Defaults to None.
-            ylabel (str, optional): Label for y axis. Defaults to None.
-            legend (str, optional): Title for legend. Defaults to None.
-
-        """
-        import matplotlib.pyplot as plt
-
-        self.plot(**kwargs)
-        plt.savefig(filename, bbox_inches='tight', dpi=dpi)
-        log.info(f"Saved 2D UMAP to {col.green(filename)}")
-
-    def save_3d(
-        self,
-        filename: str,
-        dpi: int = 300,
-        **kwargs
-
-    ):
-        """Save 3D plot of slide map.
-
-        Args:
-            filename (str): _description_
-            dpi (int, optional): _description_. Defaults to 300.
-
-        Keyword args:
-            z (list, optional): Values for z axis. Must supply z or feature.
-                Defaults to None.
-            feature (int, optional): Int, feature to plot on 3rd axis.
-                Must supply z or feature. Defaults to None.
-            subsample (int, optional): Subsample to only include this many
-                tiles on plot. Defaults to None.
-
-        """
-        import matplotlib.pyplot as plt
-
-        self.plot_3d(**kwargs)
-        plt.savefig(filename, bbox_inches='tight', dpi=dpi)
-        log.info(f"Saved 3D UMAP to {col.green(filename)}")
-
-    def save_cache(self, path: Optional[str] = None) -> None:
-        """Save cache of coordinates to PKL file.
-
-        Args:
-            path (str, optional): Save cache to this location. If None,
-                will use `self.cache`.
-        """
-        if path is None:
-            path = self.cache
-        if path:
-            self.data.to_parquet(path)
-            log.info(f"Wrote slide map cache to {col.green(path)}")
-
-    def load_cache(self, path: Optional[str] = None) -> bool:
-        """Load coordinates from PKL cache.
-
-        Args:
-            path (str, optional): Load cache from this location. If None,
-                will use `self.cache`.
-
-        Returns:
-            bool: If successfully loaded from cache.
-        """
-        if path is None:
-            path = self.cache
-        if path is None:
-            raise errors.SlideMapError("No cache set or given.")
-        try:
-            self.data = pd.read_parquet(path)
-            log.info(f"Loaded UMAP cache from {col.green(path)}")
-            return True
-        except FileNotFoundError:
-            log.info(f"No UMAP cache found at {col.green(path)}")
-        except Exception:
-            log.error(
-                f"Error loading slide map cache at {col.green(path)}, "
-                "ensure it is a valid parquet-format dataframe."
-            )
-        return False
-
-
-def _generate_tile_roc(
-    idx_and_yp: Tuple[int, np.ndarray],
-    y_true: np.ndarray,
-    data_dir: str,
-    label: str,
-    neptune_run: Optional["neptune.Run"] = None
-) -> Tuple[float, float, float]:
-    """Generate tile-level ROC. Defined separately for multiprocessing.
-
-    Args:
-        idx_and_yp (Tuple[int, np.ndarray]): Category index and y_pred.
-        y_true (np.ndarray): Array of y_true (category label). Will compare
-            against idx.
-        data_dir (str): Out directory in which to save ROC curves / histograms.
-        label (str): Label for the plots.
-        neptune_run (neptune.Run, optional): Neptune run. Defaults to None.
-
-    Returns:
-        float: AUROC
-
-        float: AP (average precision)
-
-        float: Optimal threshold (via Youden's J)
-    """
-    i, y_pred = idx_and_yp
-    y_true = (y_true == i).astype(int)
-    label = label + str(i)
-    auc, ap, thresh = generate_roc(
-        y_true,
-        y_pred,
-        data_dir,
-        label.format('ROC'),
-        neptune_run
-    )
-    return auc, ap, thresh  # ROC AUC, Average Precision, Optimal Threshold
-
-
-def filtered_prediction(
-    logits: np.ndarray,
-    filter: Optional[list] = None
-) -> int:
-    """Generates a prediction from a logits vector masked by a given filter.
-
-    Args:
-        filter (list(int)): List of logit indices to include when generating
-            a prediction. All other logits will be masked.
-
-    Returns:
-        int: index of prediction.
-    """
-    if filter is None:
-        prediction_mask = np.zeros(logits.shape, dtype=int)
-    else:
-        prediction_mask = np.ones(logits.shape, dtype=int)
-        prediction_mask[filter] = 0
-    masked_logits = np.ma.masked_array(logits, mask=prediction_mask)  # type: ignore
-    return int(np.argmax(masked_logits))
-
-
-def get_centroid_index(arr: np.ndarray) -> int:
-    """Calculate index nearest to centroid from a given 2D input array."""
-    km = KMeans(n_clusters=1).fit(arr)
-    closest, _ = pairwise_distances_argmin_min(km.cluster_centers_, arr)
-    return closest[0]
-
-
-def calculate_centroid(
-    act: Dict[str, np.ndarray]
-) -> Tuple[Dict[str, int], Dict[str, np.ndarray]]:
-    """Calcultes slide-level centroid indices for a provided activations dict.
-
-    Args:
-        activations (dict): Dict mapping slide names to ndarray of activations
-            across tiles, of shape (n_tiles, n_features)
-
-    Returns:
-        dict: Dict mapping slides to index of tile nearest to centroid
-        dict: Dict mapping slides to activations of tile nearest to centroid
-    """
-
-    optimal_indices = {}
-    centroid_activations = {}
-    for slide in act:
-        if not len(act[slide]):
-            continue
-        km = KMeans(n_clusters=1).fit(act[slide])
-        closest, _ = pairwise_distances_argmin_min(
-            km.cluster_centers_,
-            act[slide]
-        )
-        closest_index = closest[0]
-        closest_activations = act[slide][closest_index]
-        optimal_indices.update({slide: closest_index})
-        centroid_activations.update({slide: closest_activations})
-    return optimal_indices, centroid_activations
-
-
-def normalize_layout(
-    layout: np.ndarray,
-    min_percentile: int = 1,
-    max_percentile: int = 99,
-    relative_margin: float = 0.1
-) -> np.ndarray:
-    """Removes outliers and scales layout to between [0,1].
-
-    Args:
-        layout (np.ndarray): 2D array containing data to be scaled.
-        min_percentile (int, optional): Percentile for scaling. Defaults to 1.
-        max_percentile (int, optional): Percentile for scaling. Defaults to 99.
-        relative_margin (float, optional): Add an additional margin (fraction
-            of total plot width). Defaults to 0.1.
-
-    Returns:
-        np.ndarray: layout array, re-scaled and clipped.
-    """
-
-    # Compute percentiles
-    mins = np.percentile(layout, min_percentile, axis=(0))
-    maxs = np.percentile(layout, max_percentile, axis=(0))
-    # Add margins
-    mins -= relative_margin * (maxs - mins)
-    maxs += relative_margin * (maxs - mins)
-    # `clip` broadcasts, `[None]`s added only for readability
-    clipped = np.clip(layout, mins, maxs)
-    # embed within [0,1] along both axes
-    clipped -= clipped.min(axis=0)
-    clipped /= clipped.max(axis=0)
-    return clipped
-
-
-def save_histogram(
-    y_true: np.ndarray,
-    y_pred: np.ndarray,
-    outdir: str,
-    name: str = 'histogram',
-    subsample: int = 500,
-    neptune_run: Optional["neptune.Run"] = None
-) -> None:
-    """Generates histogram of y_pred, labeled by y_true, saving to outdir.
-
-    Args:
-        y_true (np.ndarray): y_true array.
-        y_pred (np.ndarray): y_pred array.
-        outdir (str): Target directory in which to save histogram.
-        name (str, optional): Label for plot. Defaults to 'histogram'.
-        subsample (int, optional): Subsample data. Defaults to 500.
-        neptune_run (neptune.Run, optional): Neptune run for saving the plot.
-            Defaults to None.
-    """
-    import seaborn as sns
-    from matplotlib import pyplot as plt
-
-    # Subsample
-    if subsample and y_pred.shape[0] > subsample:
-        idx = np.arange(y_pred.shape[0])
-        idx = np.random.choice(idx, subsample)
-        y_pred = y_pred[idx]
-        y_true = y_true[idx]
-
-    cat_false = y_pred[y_true == 0]
-    cat_true = y_pred[y_true == 1]
-    plt.clf()
-    plt.title('Tile-level Predictions')
-    plot_kw = {
-        'bins': 30,
-        'kde': True,
-        'stat': 'density',
-        'linewidth': 0
-    }
-    try:
-        sns.histplot(cat_false, color="skyblue", label="Negative", **plot_kw)
-        sns.histplot(cat_true, color="red", label="Positive", **plot_kw)
-    except np.linalg.LinAlgError:
-        log.warning("Unable to generate histogram, insufficient data")
-    plt.legend()
-    plt.savefig(os.path.join(outdir, f'{name}.png'))
-    if neptune_run:
-        neptune_run[f'results/graphs/{name}'].upload(
-            os.path.join(outdir, f'{name}.png')
-        )
-
-
-def plot_roc(
-    fpr: np.ndarray,
-    tpr: np.ndarray,
-    label: Optional[str] = None
-):
-    from matplotlib import pyplot as plt
-    plt.clf()
-    plt.title('ROC Curve')
-    plt.plot(fpr, tpr, 'b', label=label)
-    plt.legend(loc='lower right')
-    plt.plot([0, 1], [0, 1], 'r--')
-    plt.xlim([0, 1])
-    plt.ylim([0, 1])
-    plt.ylabel('TPR')
-    plt.xlabel('FPR')
-
-
-def plot_prc(
-    precision: np.ndarray,
-    recall: np.ndarray,
-    label: Optional[str] = None
-):
-    from matplotlib import pyplot as plt
-    plt.clf()
-    plt.title('Precision-Recall Curve')
-    plt.plot(precision, recall, 'b', label=label)
-    plt.legend(loc='lower right')
-    plt.plot([0, 1], [0, 1], 'r--')
-    plt.xlim([0, 1])
-    plt.ylim([0, 1])
-    plt.ylabel('Recall')
-    plt.xlabel('Precision')
-
-
-def generate_roc(
-    y_true: np.ndarray,
-    y_pred: np.ndarray,
-    save_dir: Optional[str] = None,
-    name: str = 'ROC',
-    neptune_run: Optional["neptune.Run"] = None
-) -> Tuple[float, float, float]:
-    """Generates and saves an ROC with a given set of y_true, y_pred values.
-
-    Args:
-        y_true (np.ndarray): y_true array.
-        y_pred (np.ndarray): y_pred array.
-        save_dir (str, optional): Path in which to save ROC curves.
-            Defaults to None.
-        name (str, optional): Name for plots. Defaults to 'ROC'.
-        neptune_run (neptune.Run, optional): Neptune run for saving plots.
-            Defaults to None.
-
-    Returns:
-        float:  AUROC
-
-        float:  AP (average precision)
-
-        float:  Optimal threshold (via Youden's J)
-    """
-
-    # ROC
-    fpr, tpr, threshold = metrics.roc_curve(y_true, y_pred)
-    roc_auc = metrics.auc(fpr, tpr)
-
-    # Precision recall
-    precision, recall, _ = metrics.precision_recall_curve(y_true, y_pred)
-    ap = metrics.average_precision_score(y_true, y_pred)
-
-    # Calculate optimal cutoff via maximizing Youden's J statistic
-    # (sens+spec-1, or TPR - FPR)
-    try:
-        max_youden = max(zip(tpr, fpr), key=lambda x: x[0]-x[1])
-        opt_thresh_index = list(zip(tpr, fpr)).index(max_youden)
-        opt_thresh = threshold[opt_thresh_index]
-    except Exception:
-        opt_thresh = -1
-    if save_dir:
-        from matplotlib import pyplot as plt
-        plot_roc(fpr, tpr, f'AUC = {roc_auc:.2f}')
-        plt.savefig(os.path.join(save_dir, f'{name}.png'))
-        plot_prc(precision, recall, label=f'AP = {ap:.2f}')
-        plt.savefig(os.path.join(save_dir, f'{name}-PRC.png'))
-        if neptune_run:
-            neptune_run[f'results/graphs/{name}'].upload(
-                os.path.join(save_dir, f'{name}.png')
-            )
-            neptune_run[f'results/graphs/{name}-PRC'].upload(
-                os.path.join(save_dir, f'{name}-PRC.png')
-            )
-    return roc_auc, ap, opt_thresh
-
-
-def generate_combined_roc(
-    y_true: np.ndarray,
-    y_pred: np.ndarray,
-    save_dir: str,
-    labels: List[Union[str, int]],
-    name: str = 'ROC',
-    neptune_run: Optional["neptune.Run"] = None
-) -> List[float]:
-    """Generates and saves overlapping ROCs.
-
-    Args:
-        y_true (np.ndarray): y_true array of shape = (n_curves, n_samples).
-        y_pred (np.ndarray): y_pred array of shape = (n_curves, n_samples).
-        save_dir (str, optional): Path in which to save ROC curves.
-            Defaults to None.
-        name (str, optional): Name for plots. Defaults to 'ROC'.
-        neptune_run (neptune.Run, optional): Neptune run for saving plots.
-            Defaults to None.
-
-    Returns:
-        float:  AUROC
-
-        float:  AP (average precision)
-
-        float:  Optimal threshold (via Youden's J)
-    """
-    from matplotlib import pyplot as plt
-
-    plt.clf()
-    plt.title(name)
-    colors = ('b', 'g', 'r', 'c', 'm', 'y', 'k')
-    rocs = []
-    for i, (yt, yp) in enumerate(zip(y_true, y_pred)):
-        fpr, tpr, threshold = metrics.roc_curve(yt, yp)
-        roc_auc = metrics.auc(fpr, tpr)
-        rocs += [roc_auc]
-        label = f'{labels[i]} (AUC: {roc_auc:.2f})'
-        plt.plot(fpr, tpr, colors[i % len(colors)], label=label)
-    plt.legend(loc='lower right')
-    plt.plot([0, 1], [0, 1], 'r--')
-    plt.xlim([0, 1])
-    plt.ylim([0, 1])
-    plt.ylabel('TPR')
-    plt.xlabel('FPR')
-    plt.savefig(os.path.join(save_dir, f'{name}.png'))
-    if neptune_run:
-        neptune_run[f'results/graphs/{name}'].upload(
-            os.path.join(save_dir, f'{name}.png')
-        )
-    return rocs
-
-
-def generate_scatter(
-    y_true: np.ndarray,
-    y_pred: np.ndarray,
-    data_dir: str,
-    name: str = '_plot',
-    neptune_run: Optional["neptune.Run"] = None
-) -> List[float]:
-    """Generate and save scatter plots and calculate R2 for each outcome.
-
-        Args:
-            y_true (np.ndarray): 2D array of labels. Observations are in first
-                dimension, second dim is the outcome.
-            y_pred (np.ndarray): 2D array of predictions.
-            data_dir (str): Path to directory in which to save plots.
-            name (str, optional): Label for filename. Defaults to '_plot'.
-            neptune_run (optional): Neptune Run. If provided, will upload plot.
-
-        Returns:
-            List[float]:    R squared for each outcome.
-    """
-    import seaborn as sns
-    from matplotlib import pyplot as plt
-
-    if y_true.shape != y_pred.shape:
-        m = f"Shape mismatch: y_true {y_true.shape} y_pred: {y_pred.shape}"
-        raise errors.StatsError(m)
-    if y_true.shape[0] < 2:
-        raise errors.StatsError("Only one observation provided, need >1")
-    r_squared = []
-    # Perform scatter for each outcome
-    for i in range(y_true.shape[1]):
-        # y = mx + b
-        m, b, r, p_val, err = stats.linregress(y_true[:, i], y_pred[:, i])
-        r_squared += [r ** 2]
-        p = sns.jointplot(x=y_true[:, i], y=y_pred[:, i], kind="reg")
-        p.set_axis_labels('y_true', 'y_pred')
-        plt.savefig(os.path.join(data_dir, f'Scatter{name}-{i}.png'))
-        if neptune_run:
-            neptune_run[f'results/graphs/Scatter{name}-{i}'].upload(
-                os.path.join(data_dir, f'Scatter{name}-{i}.png')
-            )
-    return r_squared
-
-
-def group_reduce(
-    df: DataFrame,
-    method: str = 'average',
-    patients: Optional[Dict[str, str]] = None
-) -> Dict[str, DataFrame]:
-    """Reduces tile-level predictions to group-level predictions.
-
-    Args:
-        df (DataFrame): Tile-level predictions.
-        method (str, optional): Reduction method for calculating
-            slide-level and patient-level predictions for categorical outcomes.
-            Either 'average' or 'proportion'. If 'average', will reduce with
-            average of each logit across tiles. If 'proportion', will convert
-            tile predictions into onehot encoding then reduce by averaging
-            these onehot values. Defaults to 'average'.
-    """
-    log.debug(f"Using reduce_method={method}")
-
-    if patients is not None:
-        df['patient'] = df['slide'].map(patients)
-        groups = ['slide', 'patient']
-    else:
-        groups = ['slide']
-
-    group_dfs = {
-        'tile': df
-    }
-    _df = df
-    if method == 'proportion':
-        outcome_names = [c[:-8] for c in df.columns if c.endswith('-y_pred0')]
-        for outcome in outcome_names:
-            y_pred_cols = [c for c in df.columns
-                           if c.startswith(f"{outcome}-y_pred")]
-            num_cat = len(y_pred_cols)
-            if num_cat != df.y_true.max()+1:
-                raise errors.StatsError(
-                    "Model predictions have a different number of outcome "
-                    f"categories ({df.y_true.max()+1}) "
-                    f"than provided annotations ({num_cat})"
-                )
-
-            outcome_pred_cat = df[y_pred_cols].values.max(1)
-            for i in range(num_cat):
-                _df[f'{outcome}-y_pred{i}'] = (_df[f'{outcome}-y_pred{i}']
-                                               == outcome_pred_cat).astype(int)
-
-    for group in groups:
-        group_dfs.update({group: _df.groupby(group, as_index=False).mean()})
-
-    return group_dfs
-
-
-def name_columns(
-    df: DataFrame,
-    model_type: str,
-    outcome_names: Optional[List[str]] = None
-):
-    _assert_model_type(model_type)
-
-    if outcome_names is None and model_type != 'cph':
-        raise ValueError("Must supply outcome names for categorical "
-                         "or linear models.")
-    if not isinstance(outcome_names, (list, tuple)):
-        outcome_names = [outcome_names]
-
-    if model_type == 'categorical':
-        # Update dataframe column names with outcome names
-        outcome_cols_to_replace = {}
-        for oi, outcome in enumerate(outcome_names):
-            outcome_cols_to_replace.update({
-                c: c.replace(f'out{oi}', outcome)
-                for c in df.columns
-                if c.startswith(f'out{oi}-')
-            })
-        df = df.rename(columns=outcome_cols_to_replace)
-
-    elif model_type == 'linear':
-        n_outcomes = len([c for c in df.columns if c.startswith('out0-y_pred')])
-        if not outcome_names:
-            outcome_names = [f"Outcome {i}" for i in range(n_outcomes)]
-        elif len(outcome_names) != n_outcomes:
-            raise errors.StatsError(
-                f"Number of outcome names {len(outcome_names)} does not "
-                f"match y_true {n_outcomes}"
-            )
-
-        # Rename columns
-        outcome_cols_to_replace = {}
-        def replace_dict(target, oi, ending_not_needed=False):
-            return {
-                c: f'{outcome}-{target}'
-                for c in df.columns
-                if c.startswith(f'out0-{target}') and (c.endswith(str(oi))
-                                                        or ending_not_needed)
-            }
-        for oi, outcome in enumerate(outcome_names):
-            outcome_cols_to_replace.update(replace_dict(
-                'y_true', oi, ending_not_needed=(len(outcome_names) == 1)
-            ))
-            outcome_cols_to_replace.update(replace_dict('y_pred', oi))
-            outcome_cols_to_replace.update(replace_dict('uncertainty', oi))
-        df = df.rename(columns=outcome_cols_to_replace)
-
-    else:
-        df = df.rename(columns={
-            'out0-y_pred0': 'time-y_pred',
-            'out0-y_pred1': 'event-y_true',
-            'out0-y_true0': 'time-y_true',
-
-        })
-    return df
-
-
-def basic_metrics(y_true: np.ndarray, y_pred: np.ndarray) -> Dict[str, float]:
-    """Generates metrics, including sensitivity, specificity, and accuracy.
-
-    Args:
-        y_true (np.ndarray): True labels.
-        y_pred (np.ndarray): Predictions.
-
-    Returns:
-        Dict[str, float]: Dict with metrics including accuracy, sensitivity,
-        specificity, precision, recall, f1_score, and kappa.
-    """
-    assert(len(y_true) == len(y_pred))
-    assert([y in (0, 1) for y in y_true])
-    assert([y in (0, 1) for y in y_pred])
-
-    TP = 0  # True positive
-    TN = 0  # True negative
-    FP = 0  # False positive
-    FN = 0  # False negative
-
-    for i, yt in enumerate(y_true):
-        yp = y_pred[i]
-        if yt == 1 and yp == 1:
-            TP += 1
-        elif yt == 1 and yp == 0:
-            FN += 1
-        elif yt == 0 and yp == 1:
-            FP += 1
-        elif yt == 0 and yp == 0:
-            TN += 1
-
-    results = {}
-    results['accuracy'] = (TP + TN) / (TP + TN + FP + FN)
-    results['sensitivity'] = TP / (TP + FN)
-    results['specificity'] = TN / (TN + FP)
-    results['precision'] = metrics.precision_score(y_true, y_pred)
-    results['recall'] = metrics.recall_score(y_true, y_pred)
-    results['f1_score'] = metrics.f1_score(y_true, y_pred)
-    results['kappa'] = metrics.cohen_kappa_score(y_true, y_pred)
-    return results
-
-
-def concordance_index(y_true: np.ndarray, y_pred: np.ndarray) -> float:
-    '''Calculates concordance index from a given y_true and y_pred.'''
-    E = y_pred[:, -1]
-    y_pred = y_pred[:, :-1]
-    y_pred = y_pred.flatten()
-    E = E.flatten()
-    y_true = y_true.flatten()
-    # Need -1 * concordance index, since these are log hazard ratios
-    y_pred = - y_pred
-    return c_index(y_true, y_pred, E)
-
-
-def df_from_pred(
-    y_true: Optional[List[Any]],
-    y_pred: List[Any],
-    y_std: Optional[List[Any]],
-    tile_to_slides: Union[List, np.ndarray],
-) -> DataFrame:
-    """Converts arrays of model predictions to a pandas dataframe.
-
-    Args:
-        y_true (list(np.ndarray)): List of y_true numpy arrays, one array for
-            each outcome. For linear outcomes, the length of the outer
-            list should be one, and the second shape dimension of the numpy
-            array should be the number of linear outcomes.
-        y_pred (list(np.ndarray)): List of y_pred numpy arrays, one array for each
-            outcome. For linear outcomes, the length of the outer
-            list should be one, and the second shape dimension of the numpy
-            array should be the number of linear outcomes.
-        y_std (list(np.ndarray)): List of uncertainty numpy arrays, formatted in
-            the same way as y_pred.
-        tile_to_slides (np.ndarray): Array of slide names for each tile. Length
-            should match the numpy arrays in y_true, y_pred, and y_std.
-
-    Returns:
-        DataFrame: _description_
-    """
-    len_err_msg = "{} must be a list of length equal to number of outcomes"
-    if y_true is not None and not isinstance(y_true, (list, tuple)):
-        raise ValueError(len_err_msg.format('y_true'))
-    if y_true is not None and not len(y_true) == len(y_pred):
-        raise ValueError('Length of y_pred and y_true must be equal')
-    if not isinstance(y_pred, (list, tuple)):
-        raise ValueError(len_err_msg.format('y_pred'))
-    if y_std is not None and not isinstance(y_std, (list, tuple)):
-        raise ValueError(len_err_msg.format('y_std'))
-    if y_std is not None and len(y_std) != len(y_pred):
-        raise ValueError('If y_std is provided, length must equal y_pred')
-
-    n_outcomes = len(y_pred)
-    series = {
-        'slide': pd.Series(tile_to_slides)
-    }
-    # Iterate through each outcome in y_pred
-    for oi in range(n_outcomes):
-        # Add y_pred columns
-        series.update({
-            f'out{oi}-y_pred{n}': y_pred[oi][:, n]
-            for n in range(y_pred[oi].shape[1])
-        })
-        # Add y_true columns
-        if y_true is not None:
-            if len(y_true[oi].shape) == 1:
-                series.update({
-                    f'out{oi}-y_true': y_true[oi]
-                })
-            else:
-                series.update({
-                    f'out{oi}-y_true{n}': y_true[oi][:, n]
-                    for n in range(y_true[oi].shape[1])
-                })
-        # Add uncertainty columns
-        if y_std is not None:
-            series.update({
-                f'out{oi}-uncertainty{n}': y_std[oi][:, n]
-                for n in range(y_std[oi].shape[1])
-            })
-    return pd.DataFrame(series)
-
-
-def _assert_model_type(model_type: str) -> None:
-    """Raises a ValueError if the model type is invalid."""
-    if model_type not in ('categorical', 'linear', 'cph'):
-        raise ValueError(f"Unrecognized model_type {model_type}, must be "
-                         "categorical, linear, or cph")
-
-
-def _merge_metrics(metrics_by_level: Dict[str, Dict]) -> Dict[str, Dict]:
-    levels = list(metrics_by_level.keys())
-    metrics = list(metrics_by_level[levels[0]].keys())
-    return {
-        metric: {
-            level: metrics_by_level[level][metric]
-            for level in levels
-        } for metric in metrics
-    }
-
-
-def categorical_metrics(
-    df: DataFrame,
-    label: str = '',
-    level: str = 'tile',
-    data_dir: str = '',
-    neptune_run: Optional["neptune.Run"] = None
-) -> Dict[str, Dict[str, float]]:
-    """Generates metrics from a set of predictions.
-
-    For multiple outcomes, y_true and y_pred are expected to be a list of
-    numpy arrays (each array corresponding to whole-dataset predictions
-    for a single outcome)
-
-    Args:
-        df (pd.DataFrame): Pandas DataFrame containing labels, predictions,
-            and optionally uncertainty, as returned by sf.stats.df_from_pred()
-        patients (dict): Dictionary mapping slidenames to patients.
-        model_type (str): Either 'linear', 'categorical', or 'cph'.
-
-    Keyword args:
-        outcome_names (list, optional): List of str, names for outcomes.
-            Defaults to None.
-        label (str, optional): Label prefix/suffix for saving.
-            Defaults to None.
-        data_dir (str, optional): Path to data directory for saving.
-            Defaults to None.
-
-    Returns:
-        Dict containing metrics.
-    """
-
-    label_start = "" if label == '' else f"{label}_"
-
-    # Detect the number of outcomes and confirm that the number of outcomes
-    # match the provided outcome names
-    outcome_names = [c[:-8] for c in df.columns if c.endswith('-y_pred0')]
-
-    all_auc = {outcome: [] for outcome in outcome_names}
-    all_ap = {outcome: [] for outcome in outcome_names}
-
-    def y_true_onehot(_df, i):
-        return (_df.y_true == i).astype(int)
-
-    def y_pred_onehot(_df, i):
-        return (_df[f'y_pred{i}'] == _df.y_pred_cat).astype(int)
-
-    # Perform analysis separately for each outcome column
-    for oi, outcome in enumerate(outcome_names):
-        outcome_cols = [c for c in df.columns if c.startswith(f'{outcome}-')]
-        # Remove the outcome name from the dataframe temporarily
-        outcome_df = df[outcome_cols].rename(
-            columns={
-                orig_col: orig_col.replace(f'{outcome}-', '', 1)
-                for orig_col in outcome_cols
-            }
-        )
-        log.info(f"Validation metrics for outcome {col.green(outcome)}:")
-        y_pred_cols = [c for c in outcome_df.columns if c.startswith('y_pred')]
-        num_cat = len(y_pred_cols)
-
-        if not len(y_pred_cols):
-            print("outcome:", outcome)
-            print("y_pred cols:", y_pred_cols)
-
-        # Convert to one-hot encoding
-        outcome_df['y_pred_cat'] = outcome_df[y_pred_cols].values.max(1)
-
-        log.debug(f"Calculating metrics with a multiprocessing pool")
-        ctx = mp.get_context('spawn')
-        p = ctx.Pool(8)
-        roc_fn = partial(
-            _generate_tile_roc,
-            y_true=outcome_df.y_true.values,
-            data_dir=data_dir,
-            label=str(label_start + outcome) + "_tile_{}",
-        )
-        try:
-            idx_and_yp = [(i, y_pred_onehot(outcome_df, i)) for i in range(num_cat)]
-            for i, (auc, ap, thresh) in enumerate(p.imap(roc_fn, idx_and_yp)):
-                all_auc[outcome] += [auc]
-                all_ap[outcome] += [ap]
-                log.info(
-                    f"{level}-level AUC (cat #{i:>2}): {auc:.3f} "
-                    f"{level}-level AP: {ap:.3f} (opt. threshold: {thresh:.3f})"
-                )
-        except ValueError as e:
-            # Occurs when predictions contain NaN
-            log.error(f'Error encountered when generating AUC: {e}')
-            all_auc[outcome] = -1
-            all_ap[outcome] = -1
-        p.close()
-
-        # Calculate tile-level accuracy.
-        # Category-level accuracy is determined by comparing
-        # one-hot predictions to one-hot y_true.
-        for i in range(num_cat):
-            try:
-                yt_in_cat =  y_true_onehot(outcome_df, i)
-                n_in_cat = yt_in_cat.sum()
-                correct = y_pred_onehot(outcome_df.loc[yt_in_cat > 0], i).sum()
-                category_accuracy = correct / n_in_cat
-                perc = category_accuracy * 100
-                log.info(f"Category {i} acc: {perc:.1f}% ({correct}/{n_in_cat})")
-            except IndexError:
-                log.warning(f"Error with category accuracy for cat # {i}")
-    return {
-        'auc': all_auc,
-        'ap': all_ap,
-    }
-
-
-def linear_metrics(
-    df: DataFrame,
-    label: str = '',
-    level: str = 'tile',
-    data_dir: str = '',
-    neptune_run: Optional["neptune.Run"] = None
-) -> Dict[str, Dict[str, float]]:
-    """Generates metrics from a set of predictions.
-
-    For multiple outcomes, y_true and y_pred are expected to be a list of
-    numpy arrays (each array corresponding to whole-dataset predictions
-    for a single outcome)
-
-    Args:
-        df (pd.DataFrame): Pandas DataFrame containing labels, predictions,
-            and optionally uncertainty, as returned by sf.stats.df_from_pred()
-        patients (dict): Dictionary mapping slidenames to patients.
-        model_type (str): Either 'linear', 'categorical', or 'cph'.
-
-    Keyword args:
-        outcome_names (list, optional): List of str, names for outcomes.
-            Defaults to None.
-        label (str, optional): Label prefix/suffix for saving.
-            Defaults to None.
-        data_dir (str, optional): Path to data directory for saving.
-            Defaults to None.
-        save_predictions (bool, optional): Save tile, slide, and patient-level
-            predictions to CSV. Defaults to True.
-        neptune_run (:class:`neptune.Run`, optional): Neptune run in which to
-            log results. Defaults to None.
-
-    Returns:
-        Dict containing metrics.
-    """
-
-    label_end = "" if label == '' else f"_{label}"
-
-    # Detect the outcome names
-    outcome_names = [c[:-7] for c in df.columns if c.endswith('-y_pred')]
-    _outcomes_by_true = [c[:-7] for c in df.columns if c.endswith('-y_true')]
-    if ((sorted(outcome_names) != sorted(_outcomes_by_true)) 
-       or not len(outcome_names)):
-        raise ValueError("Improperly formatted dataframe to linear_metrics(); "
-                         "could not detect outcome names. Ensure that "
-                         "prediction columns end in '-y_pred' and ground-truth "
-                         "columns end in '-y_true'. Try setting column names "
-                         "with slideflow.stats.name_columns(). "
-                         f"DataFrame columns: {list(df.columns)}")
-
-    # Calculate metrics
-    y_pred_cols = [f'{o}-y_pred' for o in outcome_names]
-    y_true_cols = [f'{o}-y_true' for o in outcome_names]
-    r_squared = generate_scatter(
-        df[y_true_cols].values,
-        df[y_pred_cols].values,
-        data_dir,
-        f"{label_end}_by_{level}",
-        neptune_run=neptune_run
-    )
-
-    # Show results
-    for o, r in zip(outcome_names, r_squared):
-        log.info(f"{col.green(o)}: R-squared ({level}-level): {r:.3f}")
-
-    return {
-        'r_squared': r_squared,
-    }
-
-
-def cph_metrics(
-    df: DataFrame,
-    level: str = 'tile',
-    label: str = '',
-    data_dir: str = '',
-    neptune_run: Optional["neptune.Run"] = None
-) -> Dict[str, Dict[str, float]]:
-    """Generates metrics from a set of predictions.
-
-    For multiple outcomes, y_true and y_pred are expected to be a list of
-    numpy arrays (each array corresponding to whole-dataset predictions
-    for a single outcome)
-
-    Args:
-        df (pd.DataFrame): Pandas DataFrame containing labels, predictions,
-            and optionally uncertainty, as returned by sf.stats.df_from_pred()
-        patients (dict): Dictionary mapping slidenames to patients.
-        model_type (str): Either 'linear', 'categorical', or 'cph'.
-
-    Keyword args:
-        outcome_names (list, optional): List of str, names for outcomes.
-            Defaults to None.
-        label (str, optional): Label prefix/suffix for saving.
-            Defaults to None.
-        data_dir (str, optional): Path to data directory for saving.
-            Defaults to None.
-        plot (bool, optional): Save scatterplot for linear outcomes.
-            Defaults to True.
-        neptune_run (:class:`neptune.Run`, optional): Neptune run in which to
-            log results. Defaults to None.
-
-    Returns:
-        Dict containing metrics.
-    """
-    cph_cols = ('time-y_true', 'time-y_pred', 'event-y_true')
-    if any(c not in df.columns for c in cph_cols):
-        raise ValueError(
-            "Improperly formatted dataframe to cph_metrics(), "
-            f"must have columns {cph_cols}. Got: {list(df.columns)}"
-        )
-
-    # Calculate metrics
-    try:
-        c_index = concordance_index(
-            df['time-y_true'].values,
-            df[['time-y_pred', 'event-y_true']].values,
-        )
-        log.info(f"C-index ({level}-level): {c_index:.3f}")
-    except ZeroDivisionError as e:
-        log.error(f"Error calculating concordance index: {e}")
-        c_index = -1
-    return {
-        'c_index': c_index
-    }
-
-
-def predict_from_dataset(
-    model: Union["tf.keras.Model", "torch.nn.Module"],
-    dataset: Union["tf.data.Dataset", "torch.utils.data.DataLoader"],
-    model_type: str,
-    pred_args: SimpleNamespace,
-    num_tiles: int = 0,
-    uq_n: int = 30,
-    reduce_method: str = 'average',
-    patients: Optional[Dict[str, str]] = None,
-    outcome_names: Optional[List[str]] = None
-) -> DataFrame:
-    """Generates predictions (y_pred, tile_to_slide) from model and dataset.
-
-    Args:
-        model (str): Path to PyTorch model.
-        dataset (tf.data.Dataset): PyTorch dataloader.
-        model_type (str, optional): 'categorical', 'linear', or 'cph'.
-            If multiple linear outcomes are present, y_true is stacked into a
-            single vector for each image. Defaults to 'categorical'.
-        pred_args (namespace): Namespace containing slide_input,
-            update_corrects, and update_loss functions.
-        num_tiles (int, optional): Used for progress bar with Tensorflow.
-            Defaults to 0.
-        uq_n (int, optional): Number of forward passes to perform
-            when calculating MC Dropout uncertainty. Defaults to 30.
-        reduce_method (str, optional): Reduction method for calculating
-            slide-level and patient-level predictions for categorical outcomes.
-            Either 'average' or 'proportion'. If 'average', will reduce with
-            average of each logit across tiles. If 'proportion', will convert
-            tile predictions into onehot encoding then reduce by averaging
-            these onehot values. Defaults to 'average'.
-
-    Returns:
-        pd.DataFrame
-    """
-    if model_type != 'categorical' and reduce_method != 'average':
-        raise ValueError(
-            f'Reduction method {reduce_method} incompatible with '
-            f'model_type {model_type}'
-        )
-
-    if sf.backend() == 'tensorflow':
-        from slideflow.model.tensorflow_utils import _predict_from_model
-    else:
-        from slideflow.model.torch_utils import _predict_from_model
-
-    df = _predict_from_model(
-        model,
-        dataset,
-        model_type,
-        pred_args,
-        num_tiles=num_tiles,
-        uq_n=uq_n
-    )
-    if outcome_names is not None or model_type == 'cph':
-        df = name_columns(df, model_type, outcome_names)
-    return group_reduce(df, method=reduce_method, patients=patients)
-
-
-def eval_from_dataset(
-    model: Union["tf.keras.Model", "torch.nn.Module"],
-    dataset: Union["tf.data.Dataset", "torch.utils.data.DataLoader"],
-    model_type: str,
-    pred_args: SimpleNamespace,
-    num_tiles: int = 0,
-    uq_n: int = 30,
-    reduce_method: str = 'average',
-    patients: Optional[Dict[str, str]] = None,
-    outcome_names: Optional[List[str]] = None,
-) -> Tuple[DataFrame, float, float]:
-    """Generates predictions (y_true, y_pred, tile_to_slide) and accuracy/loss
-    from a given model and dataset.
-
-    Args:
-        model (str): Path to PyTorch model.
-        dataset (tf.data.Dataset): PyTorch dataloader.
-        model_type (str, optional): 'categorical', 'linear', or 'cph'.
-            If multiple linear outcomes are present, y_true is stacked into a
-            single vector for each image. Defaults to 'categorical'.
-        pred_args (namespace): Namespace containing slide_input,
-            update_corrects, and update_loss functions.
-        num_tiles (int, optional): Used for progress bar with Tensorflow.
-            Defaults to 0.
-        uq_n (int, optional): Number of forward passes to perform
-            when calculating MC Dropout uncertainty. Defaults to 30.
-        reduce_method (str, optional): Reduction method for calculating
-            slide-level and patient-level predictions for categorical outcomes.
-            Either 'average' or 'proportion'. If 'average', will reduce with
-            average of each logit across tiles. If 'proportion', will convert
-            tile predictions into onehot encoding then reduce by averaging
-            these onehot values. Defaults to 'average'.
-
-    Returns:
-        pd.DataFrame, accuracy, loss
-    """
-    if model_type != 'categorical' and reduce_method != 'average':
-        raise ValueError(
-            f'Reduction method {reduce_method} incompatible with '
-            f'model_type {model_type}'
-        )
-    if sf.backend() == 'tensorflow':
-        from slideflow.model.tensorflow_utils import _eval_from_model
-    else:
-        from slideflow.model.torch_utils import _eval_from_model
-
-    df, acc, loss = _eval_from_model(
-        model,
-        dataset,
-        model_type,
-        pred_args,
-        num_tiles=num_tiles,
-        uq_n=uq_n
-    )
-
-    if outcome_names or model_type == 'cph':
-        df = name_columns(df, model_type, outcome_names)
-    return group_reduce(df, method=reduce_method, patients=patients), acc, loss
-
-
-def metrics_from_dataset(
-    model: Union["tf.keras.Model", "torch.nn.Module"],
-    model_type: str,
-    patients: Dict[str, str],
-    dataset: Union["tf.data.Dataset", "torch.utils.data.DataLoader"],
-    pred_args: SimpleNamespace,
-    num_tiles: int = 0,
-    outcome_names: Optional[List[str]] = None,
-    reduce_method: str = 'average',
-    label: str = '',
-    save_predictions: bool = False,
-    data_dir: str = '',
-    **kwargs
-) -> Tuple[Dict, float, float]:
-
-    """Evaluate performance of a given model on a given TFRecord dataset,
-    generating a variety of statistical outcomes and graphs.
-
-    Args:
-        model (tf.keras.Model or torch.nn.Module): Keras/Torch model to eval.
-        model_type (str): 'categorical', 'linear', or 'cph'.
-        labels (dict): Dictionary mapping slidenames to outcomes.
-        patients (dict): Dictionary mapping slidenames to patients.
-        dataset (tf.data.Dataset or torch.utils.data.DataLoader): Dataset.
-        pred_args (namespace, optional): Additional arguments to tensorflow and
-            torch backends.
-        num_tiles (int, optional): Number of total tiles expected in dataset.
-            Used for progress bar. Defaults to 0.
-
-    Keyword args:
-        reduce_method (str, optional): Reduction method for calculating
-            slide-level and patient-level predictions for categorical outcomes.
-            Either 'average' or 'proportion'. If 'average', will reduce with
-            average of each logit across tiles. If 'proportion', will convert
-            tile predictions into onehot encoding then reduce by averaging
-            these onehot values. Defaults to 'average'.
-        label (str, optional): Label prefix/suffix for saving.
-            Defaults to None.
-        outcome_names (list, optional): List of str, names for outcomes.
-            Defaults to None.
-        data_dir (str): Path to data directory for saving.
-            Defaults to empty string (current directory).
-        save_predictions (bool, optional): Save tile, slide, and patient-level
-            predictions to CSV. Defaults to True.
-        neptune_run (:class:`neptune.Run`, optional): Neptune run in which to
-            log results. Defaults to None.
-
-    Returns:
-        metrics [dict], accuracy [float], loss [float]
-    """
-    _assert_model_type(model_type)
-    dfs, acc, loss = eval_from_dataset(
-        model,
-        dataset,
-        model_type,
-        pred_args,
-        num_tiles=num_tiles,
-        patients=patients,
-        outcome_names=outcome_names,
-        reduce_method=reduce_method
-    )
-
-    # Save predictions
-    if save_predictions:
-        label_end = f'_{label}' if label else ''
-        for level, _df in dfs.items():
-            path = join(data_dir, f"{level}_predictions{label_end}")
-            if save_predictions == 'csv':
-                _df.to_csv(path+'.csv')
-            elif save_predictions == 'feather':
-                import pyarrow.feather as feather
-                feather.write_feather(_df, path+'.feather')
-            else:
-                _df.to_parquet(path+'.parquet.gzip', compression='gzip')
-
-    # Calculate metrics
-    def metrics_by_level(metrics_function):
-        return _merge_metrics({
-            level: metrics_function(
-                _df,
-                level=level,
-                data_dir=data_dir,
-                label=label,
-                **kwargs
-            ) for level, _df in dfs.items()
-        })
-
-    if model_type == 'categorical':
-        metrics = metrics_by_level(categorical_metrics)
-    elif model_type == 'linear':
-        metrics = metrics_by_level(linear_metrics)
-    else:
-        metrics = metrics_by_level(cph_metrics)
-
-    log.debug(f'Metrics generation complete.')
-    return metrics, acc, loss
