+++ slideflow/stats/metrics.py
@@ -1,0 +1,822 @@
+import multiprocessing as mp
+from functools import partial
+from os.path import join
+from types import SimpleNamespace
+from typing import TYPE_CHECKING, Any, Dict, List, Optional, Tuple, Union
+
+import numpy as np
+import pandas as pd
+from lifelines.utils import concordance_index as c_index
+from sklearn import metrics
+from pandas.core.frame import DataFrame
+
+import slideflow as sf
+from slideflow import errors
+from slideflow.util import colors as col
+from slideflow.util import log
+
+if TYPE_CHECKING:
+    import neptune.new as neptune
+    import tensorflow as tf
+    import torch
+
+
+def _assert_model_type(model_type: str) -> None:
+    """Raises a ValueError if the model type is invalid."""
+    if model_type not in ('categorical', 'linear', 'cph'):
+        raise ValueError(f"Unrecognized model_type {model_type}, must be "
+                         "categorical, linear, or cph")
+
+
+def _generate_tile_roc(
+    idx_and_yp: Tuple[int, np.ndarray],
+    y_true: np.ndarray,
+    data_dir: str,
+    label: str,
+    neptune_run: Optional["neptune.Run"] = None
+) -> Tuple[float, float, float]:
+    """Generate tile-level ROC. Defined separately for multiprocessing.
+
+    Args:
+        idx_and_yp (Tuple[int, np.ndarray]): Category index and y_pred.
+        y_true (np.ndarray): Array of y_true (category label). Will compare
+            against idx.
+        data_dir (str): Out directory in which to save ROC curves / histograms.
+        label (str): Label for the plots.
+        neptune_run (neptune.Run, optional): Neptune run. Defaults to None.
+
+    Returns:
+        float: AUROC
+
+        float: AP (average precision)
+
+        float: Optimal threshold (via Youden's J)
+    """
+    i, y_pred = idx_and_yp
+    y_true = (y_true == i).astype(int)
+    label = label + str(i)
+    auc, ap, thresh = generate_roc(
+        y_true,
+        y_pred,
+        data_dir,
+        label.format('ROC'),
+        neptune_run
+    )
+    return auc, ap, thresh  # ROC AUC, Average Precision, Optimal Threshold
+
+
+def _merge_metrics(metrics_by_level: Dict[str, Dict]) -> Dict[str, Dict]:
+    levels = list(metrics_by_level.keys())
+    metrics = list(metrics_by_level[levels[0]].keys())
+    return {
+        metric: {
+            level: metrics_by_level[level][metric]
+            for level in levels
+        } for metric in metrics
+    }
+
+
+def basic_metrics(y_true: np.ndarray, y_pred: np.ndarray) -> Dict[str, float]:
+    """Generates metrics, including sensitivity, specificity, and accuracy.
+
+    Args:
+        y_true (np.ndarray): True labels.
+        y_pred (np.ndarray): Predictions.
+
+    Returns:
+        Dict[str, float]: Dict with metrics including accuracy, sensitivity,
+        specificity, precision, recall, f1_score, and kappa.
+    """
+    assert(len(y_true) == len(y_pred))
+    assert([y in (0, 1) for y in y_true])
+    assert([y in (0, 1) for y in y_pred])
+
+    TP = 0  # True positive
+    TN = 0  # True negative
+    FP = 0  # False positive
+    FN = 0  # False negative
+
+    for i, yt in enumerate(y_true):
+        yp = y_pred[i]
+        if yt == 1 and yp == 1:
+            TP += 1
+        elif yt == 1 and yp == 0:
+            FN += 1
+        elif yt == 0 and yp == 1:
+            FP += 1
+        elif yt == 0 and yp == 0:
+            TN += 1
+
+    results = {}
+    results['accuracy'] = (TP + TN) / (TP + TN + FP + FN)
+    results['sensitivity'] = TP / (TP + FN)
+    results['specificity'] = TN / (TN + FP)
+    results['precision'] = metrics.precision_score(y_true, y_pred)
+    results['recall'] = metrics.recall_score(y_true, y_pred)
+    results['f1_score'] = metrics.f1_score(y_true, y_pred)
+    results['kappa'] = metrics.cohen_kappa_score(y_true, y_pred)
+    return results
+
+
+def categorical_metrics(
+    df: DataFrame,
+    label: str = '',
+    level: str = 'tile',
+    data_dir: str = '',
+    neptune_run: Optional["neptune.Run"] = None
+) -> Dict[str, Dict[str, float]]:
+    """Generates metrics from a set of predictions.
+
+    For multiple outcomes, y_true and y_pred are expected to be a list of
+    numpy arrays (each array corresponding to whole-dataset predictions
+    for a single outcome)
+
+    Args:
+        df (pd.DataFrame): Pandas DataFrame containing labels, predictions,
+            and optionally uncertainty, as returned by sf.stats.df_from_pred()
+        patients (dict): Dictionary mapping slidenames to patients.
+        model_type (str): Either 'linear', 'categorical', or 'cph'.
+
+    Keyword args:
+        outcome_names (list, optional): List of str, names for outcomes.
+            Defaults to None.
+        label (str, optional): Label prefix/suffix for saving.
+            Defaults to None.
+        data_dir (str, optional): Path to data directory for saving.
+            Defaults to None.
+
+    Returns:
+        Dict containing metrics.
+    """
+
+    label_start = "" if label == '' else f"{label}_"
+
+    # Detect the number of outcomes and confirm that the number of outcomes
+    # match the provided outcome names
+    outcome_names = [c[:-8] for c in df.columns if c.endswith('-y_pred0')]
+
+    all_auc = {outcome: [] for outcome in outcome_names}
+    all_ap = {outcome: [] for outcome in outcome_names}
+
+    def y_true_onehot(_df, i):
+        return (_df.y_true == i).astype(int)
+
+    def y_pred_onehot(_df, i):
+        return (_df[f'y_pred{i}'] == _df.y_pred_cat).astype(int)
+
+    # Perform analysis separately for each outcome column
+    for oi, outcome in enumerate(outcome_names):
+        outcome_cols = [c for c in df.columns if c.startswith(f'{outcome}-')]
+        # Remove the outcome name from the dataframe temporarily
+        outcome_df = df[outcome_cols].rename(
+            columns={
+                orig_col: orig_col.replace(f'{outcome}-', '', 1)
+                for orig_col in outcome_cols
+            }
+        )
+        log.info(f"Validation metrics for outcome {col.green(outcome)}:")
+        y_pred_cols = [c for c in outcome_df.columns if c.startswith('y_pred')]
+        num_cat = len(y_pred_cols)
+
+        if not len(y_pred_cols):
+            print("outcome:", outcome)
+            print("y_pred cols:", y_pred_cols)
+
+        # Convert to one-hot encoding
+        outcome_df['y_pred_cat'] = outcome_df[y_pred_cols].values.max(1)
+
+        log.debug(f"Calculating metrics with a multiprocessing pool")
+        ctx = mp.get_context('spawn')
+        p = ctx.Pool(8)
+        roc_fn = partial(
+            _generate_tile_roc,
+            y_true=outcome_df.y_true.values,
+            data_dir=data_dir,
+            label=str(label_start + outcome) + "_tile_{}",
+        )
+        try:
+            idx_and_yp = [(i, y_pred_onehot(outcome_df, i)) for i in range(num_cat)]
+            for i, (auc, ap, thresh) in enumerate(p.imap(roc_fn, idx_and_yp)):
+                all_auc[outcome] += [auc]
+                all_ap[outcome] += [ap]
+                log.info(
+                    f"{level}-level AUC (cat #{i:>2}): {auc:.3f} "
+                    f"{level}-level AP: {ap:.3f} (opt. threshold: {thresh:.3f})"
+                )
+        except ValueError as e:
+            # Occurs when predictions contain NaN
+            log.error(f'Error encountered when generating AUC: {e}')
+            all_auc[outcome] = -1
+            all_ap[outcome] = -1
+        p.close()
+
+        # Calculate tile-level accuracy.
+        # Category-level accuracy is determined by comparing
+        # one-hot predictions to one-hot y_true.
+        for i in range(num_cat):
+            try:
+                yt_in_cat =  y_true_onehot(outcome_df, i)
+                n_in_cat = yt_in_cat.sum()
+                correct = y_pred_onehot(outcome_df.loc[yt_in_cat > 0], i).sum()
+                category_accuracy = correct / n_in_cat
+                perc = category_accuracy * 100
+                log.info(f"Category {i} acc: {perc:.1f}% ({correct}/{n_in_cat})")
+            except IndexError:
+                log.warning(f"Error with category accuracy for cat # {i}")
+    return {
+        'auc': all_auc,
+        'ap': all_ap,
+    }
+
+
+def concordance_index(y_true: np.ndarray, y_pred: np.ndarray) -> float:
+    '''Calculates concordance index from a given y_true and y_pred.'''
+    E = y_pred[:, -1]
+    y_pred = y_pred[:, :-1]
+    y_pred = y_pred.flatten()
+    E = E.flatten()
+    y_true = y_true.flatten()
+    # Need -1 * concordance index, since these are log hazard ratios
+    y_pred = - y_pred
+    return c_index(y_true, y_pred, E)
+
+
+def cph_metrics(
+    df: DataFrame,
+    level: str = 'tile',
+    label: str = '',
+    data_dir: str = '',
+    neptune_run: Optional["neptune.Run"] = None
+) -> Dict[str, Dict[str, float]]:
+    """Generates metrics from a set of predictions.
+
+    For multiple outcomes, y_true and y_pred are expected to be a list of
+    numpy arrays (each array corresponding to whole-dataset predictions
+    for a single outcome)
+
+    Args:
+        df (pd.DataFrame): Pandas DataFrame containing labels, predictions,
+            and optionally uncertainty, as returned by sf.stats.df_from_pred()
+        patients (dict): Dictionary mapping slidenames to patients.
+        model_type (str): Either 'linear', 'categorical', or 'cph'.
+
+    Keyword args:
+        outcome_names (list, optional): List of str, names for outcomes.
+            Defaults to None.
+        label (str, optional): Label prefix/suffix for saving.
+            Defaults to None.
+        data_dir (str, optional): Path to data directory for saving.
+            Defaults to None.
+        plot (bool, optional): Save scatterplot for linear outcomes.
+            Defaults to True.
+        neptune_run (:class:`neptune.Run`, optional): Neptune run in which to
+            log results. Defaults to None.
+
+    Returns:
+        Dict containing metrics.
+    """
+    cph_cols = ('time-y_true', 'time-y_pred', 'event-y_true')
+    if any(c not in df.columns for c in cph_cols):
+        raise ValueError(
+            "Improperly formatted dataframe to cph_metrics(), "
+            f"must have columns {cph_cols}. Got: {list(df.columns)}"
+        )
+
+    # Calculate metrics
+    try:
+        c_index = concordance_index(
+            df['time-y_true'].values,
+            df[['time-y_pred', 'event-y_true']].values,
+        )
+        log.info(f"C-index ({level}-level): {c_index:.3f}")
+    except ZeroDivisionError as e:
+        log.error(f"Error calculating concordance index: {e}")
+        c_index = -1
+    return {
+        'c_index': c_index
+    }
+
+
+def df_from_pred(
+    y_true: Optional[List[Any]],
+    y_pred: List[Any],
+    y_std: Optional[List[Any]],
+    tile_to_slides: Union[List, np.ndarray],
+) -> DataFrame:
+    """Converts arrays of model predictions to a pandas dataframe.
+
+    Args:
+        y_true (list(np.ndarray)): List of y_true numpy arrays, one array for
+            each outcome. For linear outcomes, the length of the outer
+            list should be one, and the second shape dimension of the numpy
+            array should be the number of linear outcomes.
+        y_pred (list(np.ndarray)): List of y_pred numpy arrays, one array for each
+            outcome. For linear outcomes, the length of the outer
+            list should be one, and the second shape dimension of the numpy
+            array should be the number of linear outcomes.
+        y_std (list(np.ndarray)): List of uncertainty numpy arrays, formatted in
+            the same way as y_pred.
+        tile_to_slides (np.ndarray): Array of slide names for each tile. Length
+            should match the numpy arrays in y_true, y_pred, and y_std.
+
+    Returns:
+        DataFrame: _description_
+    """
+    len_err_msg = "{} must be a list of length equal to number of outcomes"
+    if y_true is not None and not isinstance(y_true, (list, tuple)):
+        raise ValueError(len_err_msg.format('y_true'))
+    if y_true is not None and not len(y_true) == len(y_pred):
+        raise ValueError('Length of y_pred and y_true must be equal')
+    if not isinstance(y_pred, (list, tuple)):
+        raise ValueError(len_err_msg.format('y_pred'))
+    if y_std is not None and not isinstance(y_std, (list, tuple)):
+        raise ValueError(len_err_msg.format('y_std'))
+    if y_std is not None and len(y_std) != len(y_pred):
+        raise ValueError('If y_std is provided, length must equal y_pred')
+
+    n_outcomes = len(y_pred)
+    series = {
+        'slide': pd.Series(tile_to_slides)
+    }
+    # Iterate through each outcome in y_pred
+    for oi in range(n_outcomes):
+        # Add y_pred columns
+        series.update({
+            f'out{oi}-y_pred{n}': y_pred[oi][:, n]
+            for n in range(y_pred[oi].shape[1])
+        })
+        # Add y_true columns
+        if y_true is not None:
+            if len(y_true[oi].shape) == 1:
+                series.update({
+                    f'out{oi}-y_true': y_true[oi]
+                })
+            else:
+                series.update({
+                    f'out{oi}-y_true{n}': y_true[oi][:, n]
+                    for n in range(y_true[oi].shape[1])
+                })
+        # Add uncertainty columns
+        if y_std is not None:
+            series.update({
+                f'out{oi}-uncertainty{n}': y_std[oi][:, n]
+                for n in range(y_std[oi].shape[1])
+            })
+    return pd.DataFrame(series)
+
+
+def eval_from_dataset(
+    model: Union["tf.keras.Model", "torch.nn.Module"],
+    dataset: Union["tf.data.Dataset", "torch.utils.data.DataLoader"],
+    model_type: str,
+    pred_args: SimpleNamespace,
+    num_tiles: int = 0,
+    uq_n: int = 30,
+    reduce_method: str = 'average',
+    patients: Optional[Dict[str, str]] = None,
+    outcome_names: Optional[List[str]] = None,
+) -> Tuple[DataFrame, float, float]:
+    """Generates predictions (y_true, y_pred, tile_to_slide) and accuracy/loss
+    from a given model and dataset.
+
+    Args:
+        model (str): Path to PyTorch model.
+        dataset (tf.data.Dataset): PyTorch dataloader.
+        model_type (str, optional): 'categorical', 'linear', or 'cph'.
+            If multiple linear outcomes are present, y_true is stacked into a
+            single vector for each image. Defaults to 'categorical'.
+        pred_args (namespace): Namespace containing slide_input,
+            update_corrects, and update_loss functions.
+        num_tiles (int, optional): Used for progress bar with Tensorflow.
+            Defaults to 0.
+        uq_n (int, optional): Number of forward passes to perform
+            when calculating MC Dropout uncertainty. Defaults to 30.
+        reduce_method (str, optional): Reduction method for calculating
+            slide-level and patient-level predictions for categorical outcomes.
+            Either 'average' or 'proportion'. If 'average', will reduce with
+            average of each logit across tiles. If 'proportion', will convert
+            tile predictions into onehot encoding then reduce by averaging
+            these onehot values. Defaults to 'average'.
+
+    Returns:
+        pd.DataFrame, accuracy, loss
+    """
+    if model_type != 'categorical' and reduce_method != 'average':
+        raise ValueError(
+            f'Reduction method {reduce_method} incompatible with '
+            f'model_type {model_type}'
+        )
+    if sf.backend() == 'tensorflow':
+        from slideflow.model.tensorflow_utils import _eval_from_model
+    else:
+        from slideflow.model.torch_utils import _eval_from_model
+
+    df, acc, loss = _eval_from_model(
+        model,
+        dataset,
+        model_type,
+        pred_args,
+        num_tiles=num_tiles,
+        uq_n=uq_n
+    )
+
+    if outcome_names or model_type == 'cph':
+        df = name_columns(df, model_type, outcome_names)
+    return group_reduce(df, method=reduce_method, patients=patients), acc, loss
+
+
+def generate_roc(
+    y_true: np.ndarray,
+    y_pred: np.ndarray,
+    save_dir: Optional[str] = None,
+    name: str = 'ROC',
+    neptune_run: Optional["neptune.Run"] = None
+) -> Tuple[float, float, float]:
+    """Generates and saves an ROC with a given set of y_true, y_pred values.
+
+    Args:
+        y_true (np.ndarray): y_true array.
+        y_pred (np.ndarray): y_pred array.
+        save_dir (str, optional): Path in which to save ROC curves.
+            Defaults to None.
+        name (str, optional): Name for plots. Defaults to 'ROC'.
+        neptune_run (neptune.Run, optional): Neptune run for saving plots.
+            Defaults to None.
+
+    Returns:
+        float:  AUROC
+
+        float:  AP (average precision)
+
+        float:  Optimal threshold (via Youden's J)
+    """
+
+    # ROC
+    fpr, tpr, threshold = metrics.roc_curve(y_true, y_pred)
+    roc_auc = metrics.auc(fpr, tpr)
+
+    # Precision recall
+    precision, recall, _ = metrics.precision_recall_curve(y_true, y_pred)
+    ap = metrics.average_precision_score(y_true, y_pred)
+
+    # Calculate optimal cutoff via maximizing Youden's J statistic
+    # (sens+spec-1, or TPR - FPR)
+    try:
+        max_youden = max(zip(tpr, fpr), key=lambda x: x[0]-x[1])
+        opt_thresh_index = list(zip(tpr, fpr)).index(max_youden)
+        opt_thresh = threshold[opt_thresh_index]
+    except Exception:
+        opt_thresh = -1
+    if save_dir:
+        from matplotlib import pyplot as plt
+        sf.stats.plot.roc(fpr, tpr, f'AUC = {roc_auc:.2f}')
+        plt.savefig(join(save_dir, f'{name}.png'))
+        sf.stats.plot.prc(precision, recall, label=f'AP = {ap:.2f}')
+        plt.savefig(join(save_dir, f'{name}-PRC.png'))
+        if neptune_run:
+            neptune_run[f'results/graphs/{name}'].upload(
+                join(save_dir, f'{name}.png')
+            )
+            neptune_run[f'results/graphs/{name}-PRC'].upload(
+                join(save_dir, f'{name}-PRC.png')
+            )
+    return roc_auc, ap, opt_thresh
+
+
+def group_reduce(
+    df: DataFrame,
+    method: str = 'average',
+    patients: Optional[Dict[str, str]] = None
+) -> Dict[str, DataFrame]:
+    """Reduces tile-level predictions to group-level predictions.
+
+    Args:
+        df (DataFrame): Tile-level predictions.
+        method (str, optional): Reduction method for calculating
+            slide-level and patient-level predictions for categorical outcomes.
+            Either 'average' or 'proportion'. If 'average', will reduce with
+            average of each logit across tiles. If 'proportion', will convert
+            tile predictions into onehot encoding then reduce by averaging
+            these onehot values. Defaults to 'average'.
+    """
+    log.debug(f"Using reduce_method={method}")
+
+    if patients is not None:
+        df['patient'] = df['slide'].map(patients)
+        groups = ['slide', 'patient']
+    else:
+        groups = ['slide']
+
+    group_dfs = {
+        'tile': df
+    }
+    _df = df
+    if method == 'proportion':
+        outcome_names = [c[:-8] for c in df.columns if c.endswith('-y_pred0')]
+        for outcome in outcome_names:
+            y_pred_cols = [c for c in df.columns
+                           if c.startswith(f"{outcome}-y_pred")]
+            num_cat = len(y_pred_cols)
+            if num_cat != df.y_true.max()+1:
+                raise errors.StatsError(
+                    "Model predictions have a different number of outcome "
+                    f"categories ({df.y_true.max()+1}) "
+                    f"than provided annotations ({num_cat})"
+                )
+
+            outcome_pred_cat = df[y_pred_cols].values.max(1)
+            for i in range(num_cat):
+                _df[f'{outcome}-y_pred{i}'] = (_df[f'{outcome}-y_pred{i}']
+                                               == outcome_pred_cat).astype(int)
+
+    for group in groups:
+        group_dfs.update({group: _df.groupby(group, as_index=False).mean()})
+
+    return group_dfs
+
+
+def linear_metrics(
+    df: DataFrame,
+    label: str = '',
+    level: str = 'tile',
+    data_dir: str = '',
+    neptune_run: Optional["neptune.Run"] = None
+) -> Dict[str, Dict[str, float]]:
+    """Generates metrics from a set of predictions.
+
+    For multiple outcomes, y_true and y_pred are expected to be a list of
+    numpy arrays (each array corresponding to whole-dataset predictions
+    for a single outcome)
+
+    Args:
+        df (pd.DataFrame): Pandas DataFrame containing labels, predictions,
+            and optionally uncertainty, as returned by sf.stats.df_from_pred()
+        patients (dict): Dictionary mapping slidenames to patients.
+        model_type (str): Either 'linear', 'categorical', or 'cph'.
+
+    Keyword args:
+        outcome_names (list, optional): List of str, names for outcomes.
+            Defaults to None.
+        label (str, optional): Label prefix/suffix for saving.
+            Defaults to None.
+        data_dir (str, optional): Path to data directory for saving.
+            Defaults to None.
+        save_predictions (bool, optional): Save tile, slide, and patient-level
+            predictions to CSV. Defaults to True.
+        neptune_run (:class:`neptune.Run`, optional): Neptune run in which to
+            log results. Defaults to None.
+
+    Returns:
+        Dict containing metrics.
+    """
+
+    label_end = "" if label == '' else f"_{label}"
+
+    # Detect the outcome names
+    outcome_names = [c[:-7] for c in df.columns if c.endswith('-y_pred')]
+    _outcomes_by_true = [c[:-7] for c in df.columns if c.endswith('-y_true')]
+    if ((sorted(outcome_names) != sorted(_outcomes_by_true)) 
+       or not len(outcome_names)):
+        raise ValueError("Improperly formatted dataframe to linear_metrics(); "
+                         "could not detect outcome names. Ensure that "
+                         "prediction columns end in '-y_pred' and ground-truth "
+                         "columns end in '-y_true'. Try setting column names "
+                         "with slideflow.stats.name_columns(). "
+                         f"DataFrame columns: {list(df.columns)}")
+
+    # Calculate metrics
+    y_pred_cols = [f'{o}-y_pred' for o in outcome_names]
+    y_true_cols = [f'{o}-y_true' for o in outcome_names]
+    r_squared = sf.stats.plot.scatter(
+        df[y_true_cols].values,
+        df[y_pred_cols].values,
+        data_dir,
+        f"{label_end}_by_{level}",
+        neptune_run=neptune_run
+    )
+
+    # Show results
+    for o, r in zip(outcome_names, r_squared):
+        log.info(f"{col.green(o)}: R-squared ({level}-level): {r:.3f}")
+
+    return {
+        'r_squared': r_squared,
+    }
+
+
+def metrics_from_dataset(
+    model: Union["tf.keras.Model", "torch.nn.Module"],
+    model_type: str,
+    patients: Dict[str, str],
+    dataset: Union["tf.data.Dataset", "torch.utils.data.DataLoader"],
+    pred_args: SimpleNamespace,
+    num_tiles: int = 0,
+    outcome_names: Optional[List[str]] = None,
+    reduce_method: str = 'average',
+    label: str = '',
+    save_predictions: bool = False,
+    data_dir: str = '',
+    **kwargs
+) -> Tuple[Dict, float, float]:
+
+    """Evaluate performance of a given model on a given TFRecord dataset,
+    generating a variety of statistical outcomes and graphs.
+
+    Args:
+        model (tf.keras.Model or torch.nn.Module): Keras/Torch model to eval.
+        model_type (str): 'categorical', 'linear', or 'cph'.
+        labels (dict): Dictionary mapping slidenames to outcomes.
+        patients (dict): Dictionary mapping slidenames to patients.
+        dataset (tf.data.Dataset or torch.utils.data.DataLoader): Dataset.
+        pred_args (namespace, optional): Additional arguments to tensorflow and
+            torch backends.
+        num_tiles (int, optional): Number of total tiles expected in dataset.
+            Used for progress bar. Defaults to 0.
+
+    Keyword args:
+        reduce_method (str, optional): Reduction method for calculating
+            slide-level and patient-level predictions for categorical outcomes.
+            Either 'average' or 'proportion'. If 'average', will reduce with
+            average of each logit across tiles. If 'proportion', will convert
+            tile predictions into onehot encoding then reduce by averaging
+            these onehot values. Defaults to 'average'.
+        label (str, optional): Label prefix/suffix for saving.
+            Defaults to None.
+        outcome_names (list, optional): List of str, names for outcomes.
+            Defaults to None.
+        data_dir (str): Path to data directory for saving.
+            Defaults to empty string (current directory).
+        save_predictions (bool, optional): Save tile, slide, and patient-level
+            predictions to CSV. Defaults to True.
+        neptune_run (:class:`neptune.Run`, optional): Neptune run in which to
+            log results. Defaults to None.
+
+    Returns:
+        metrics [dict], accuracy [float], loss [float]
+    """
+    _assert_model_type(model_type)
+    dfs, acc, loss = eval_from_dataset(
+        model,
+        dataset,
+        model_type,
+        pred_args,
+        num_tiles=num_tiles,
+        patients=patients,
+        outcome_names=outcome_names,
+        reduce_method=reduce_method
+    )
+
+    # Save predictions
+    if save_predictions:
+        label_end = f'_{label}' if label else ''
+        for level, _df in dfs.items():
+            path = join(data_dir, f"{level}_predictions{label_end}")
+            if save_predictions == 'csv':
+                _df.to_csv(path+'.csv')
+            elif save_predictions == 'feather':
+                import pyarrow.feather as feather
+                feather.write_feather(_df, path+'.feather')
+            else:
+                _df.to_parquet(path+'.parquet.gzip', compression='gzip')
+
+    # Calculate metrics
+    def metrics_by_level(metrics_function):
+        return _merge_metrics({
+            level: metrics_function(
+                _df,
+                level=level,
+                data_dir=data_dir,
+                label=label,
+                **kwargs
+            ) for level, _df in dfs.items()
+        })
+
+    if model_type == 'categorical':
+        metrics = metrics_by_level(categorical_metrics)
+    elif model_type == 'linear':
+        metrics = metrics_by_level(linear_metrics)
+    else:
+        metrics = metrics_by_level(cph_metrics)
+
+    log.debug(f'Metrics generation complete.')
+    return metrics, acc, loss
+
+
+def name_columns(
+    df: DataFrame,
+    model_type: str,
+    outcome_names: Optional[List[str]] = None
+):
+    _assert_model_type(model_type)
+
+    if outcome_names is None and model_type != 'cph':
+        raise ValueError("Must supply outcome names for categorical "
+                         "or linear models.")
+    if not isinstance(outcome_names, (list, tuple)):
+        outcome_names = [outcome_names]
+
+    if model_type == 'categorical':
+        # Update dataframe column names with outcome names
+        outcome_cols_to_replace = {}
+        for oi, outcome in enumerate(outcome_names):
+            outcome_cols_to_replace.update({
+                c: c.replace(f'out{oi}', outcome)
+                for c in df.columns
+                if c.startswith(f'out{oi}-')
+            })
+        df = df.rename(columns=outcome_cols_to_replace)
+
+    elif model_type == 'linear':
+        n_outcomes = len([c for c in df.columns if c.startswith('out0-y_pred')])
+        if not outcome_names:
+            outcome_names = [f"Outcome {i}" for i in range(n_outcomes)]
+        elif len(outcome_names) != n_outcomes:
+            raise errors.StatsError(
+                f"Number of outcome names {len(outcome_names)} does not "
+                f"match y_true {n_outcomes}"
+            )
+
+        # Rename columns
+        outcome_cols_to_replace = {}
+        def replace_dict(target, oi, ending_not_needed=False):
+            return {
+                c: f'{outcome}-{target}'
+                for c in df.columns
+                if c.startswith(f'out0-{target}') and (c.endswith(str(oi))
+                                                        or ending_not_needed)
+            }
+        for oi, outcome in enumerate(outcome_names):
+            outcome_cols_to_replace.update(replace_dict(
+                'y_true', oi, ending_not_needed=(len(outcome_names) == 1)
+            ))
+            outcome_cols_to_replace.update(replace_dict('y_pred', oi))
+            outcome_cols_to_replace.update(replace_dict('uncertainty', oi))
+        df = df.rename(columns=outcome_cols_to_replace)
+
+    else:
+        df = df.rename(columns={
+            'out0-y_pred0': 'time-y_pred',
+            'out0-y_pred1': 'event-y_true',
+            'out0-y_true0': 'time-y_true',
+
+        })
+    return df
+
+
+def predict_from_dataset(
+    model: Union["tf.keras.Model", "torch.nn.Module"],
+    dataset: Union["tf.data.Dataset", "torch.utils.data.DataLoader"],
+    model_type: str,
+    pred_args: SimpleNamespace,
+    num_tiles: int = 0,
+    uq_n: int = 30,
+    reduce_method: str = 'average',
+    patients: Optional[Dict[str, str]] = None,
+    outcome_names: Optional[List[str]] = None
+) -> DataFrame:
+    """Generates predictions (y_pred, tile_to_slide) from model and dataset.
+
+    Args:
+        model (str): Path to PyTorch model.
+        dataset (tf.data.Dataset): PyTorch dataloader.
+        model_type (str, optional): 'categorical', 'linear', or 'cph'.
+            If multiple linear outcomes are present, y_true is stacked into a
+            single vector for each image. Defaults to 'categorical'.
+        pred_args (namespace): Namespace containing slide_input,
+            update_corrects, and update_loss functions.
+        num_tiles (int, optional): Used for progress bar with Tensorflow.
+            Defaults to 0.
+        uq_n (int, optional): Number of forward passes to perform
+            when calculating MC Dropout uncertainty. Defaults to 30.
+        reduce_method (str, optional): Reduction method for calculating
+            slide-level and patient-level predictions for categorical outcomes.
+            Either 'average' or 'proportion'. If 'average', will reduce with
+            average of each logit across tiles. If 'proportion', will convert
+            tile predictions into onehot encoding then reduce by averaging
+            these onehot values. Defaults to 'average'.
+
+    Returns:
+        pd.DataFrame
+    """
+    if model_type != 'categorical' and reduce_method != 'average':
+        raise ValueError(
+            f'Reduction method {reduce_method} incompatible with '
+            f'model_type {model_type}'
+        )
+
+    if sf.backend() == 'tensorflow':
+        from slideflow.model.tensorflow_utils import _predict_from_model
+    else:
+        from slideflow.model.torch_utils import _predict_from_model
+
+    df = _predict_from_model(
+        model,
+        dataset,
+        model_type,
+        pred_args,
+        num_tiles=num_tiles,
+        uq_n=uq_n
+    )
+    if outcome_names is not None or model_type == 'cph':
+        df = name_columns(df, model_type, outcome_names)
+    return group_reduce(df, method=reduce_method, patients=patients)
